refactor: move clickable substitution component definition into new file for better unit testing
