+++ medieval-recipe-app/src/translation/Translation.jsx
@@ -1,31 +1,6 @@
 import PropTypes from 'prop-types';
-import { useState } from 'react';
 
-import GlossaryEntry from './GlossaryEntry.jsx';
-
-function ClickableSubstitution({ word, meanings, setGlossaryEntry }) {
-    ClickableSubstitution.propTypes = {
-        word: PropTypes.string,
-        meanings: PropTypes.arrayOf(PropTypes.object),
-        setGlossaryEntry: PropTypes.func,
-    }
-    const defaultSubstitution = meanings[0] && meanings[0].substitution ? meanings[0].substitution : word
-    const [substitution, setSubstitution] = useState(defaultSubstitution);
-
-    const handleClick = () => {
-        setGlossaryEntry(<GlossaryEntry
-            word={word}
-            meanings={meanings}
-            setSubstitution={setSubstitution}
-        />);
-    }
-
-    return (
-        <a href='#' onClick={(e) => { e.preventDefault(); handleClick(); }} >
-            {substitution}
-        </a>
-    );
-}
+import ClickableSubstitution from './ClickableSubstitution.jsx';
 
 export default function Translation({ now, translationArray, setGlossaryEntry }) {
     Translation.propTypes = {
@@ -35,7 +10,7 @@ export default function Translation({ now, translationArray, setGlossaryEntry })
         setGlossaryEntry: PropTypes.func,
     }
 
-    const gridItems = translationArray.map((x, i) => {
+    const paragraphContents = translationArray.map((x, i) => {
         if (typeof x === "string") {
             return x;
         } else {
@@ -53,7 +28,7 @@ export default function Translation({ now, translationArray, setGlossaryEntry })
 
     return (
         <p>
-            {gridItems}
+            {paragraphContents}
         </p>
     );
 }
